Implementing charcrea

+++ Seeds of Apocrypha/src/Entities/UI/Slider.h
@@ -17,7 +17,7 @@ private:
     //Knob stuff
     u_ptr<Sprite> knob_spr;
     u_ptr<Text> knob_label;
-    int knob_pos_max = 0;
-    int knob_pos_min = 0;
-    int knob_pos = 0;
+    float knob_pos_max = 0;
+    float knob_pos_min = 0;
+    float knob_pos = 0;
 };
